# docs/docs/api_and_sdks/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Philter API",
    "version": "4.0.0-SNAPSHOT"
  },
  "servers": [
    {
      "description": "Philter API",
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/documents/{documentId}/status": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Get the status of an async redaction.",
        "description": "Requires the `documents:read` scope.",
        "operationId": "getStatus",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "documentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "owner",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetRedactionStatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient scope, disabled API, or access denied."
          },
          "404": {
            "description": "Unknown document id.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "406": {
            "description": "No representation matches Accept."
          },
          "413": {
            "description": "Content Too Large",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Media Type",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GetRedactionStatusResponse": {
        "properties": {
          "documentId": {
            "type": "string"
          },
          "effectiveConfigurationHash": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "description": "A Philter API key, sent as `Authorization: Bearer <api key>`. Create one on the API Keys page of the dashboard.",
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "apiKey": []
    }
  ],
...
}